requirements.txt
fake-bpy-module-latest==20251003
